#219 découpage du modèle
