+++ base/models/attribution.py
@@ -23,25 +23,21 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-import string
-from django.conf.urls import include, url
-from django.contrib import admin
-import random
-import re
+from django.db import models
 
 
-def admnin_page_url() :
-    return ''.join(random.choice(string.ascii_letters + string.digits) for _ in range(random.randint(10, 20)))
+class Attribution(models.Model):
+    FUNCTION_CHOICES = (
+        ('COORDINATOR','Coordinator'),
+        ('PROFESSOR','Professor'))
 
-urlpatterns = [
-    url(r'^'+re.escape(admnin_page_url())+r'/', admin.site.urls),
-    url(r'', include('base.urls')),
-    url(r'', include('internship.urls')),
-]
+    external_id   = models.CharField(max_length=100, blank=True, null=True)
+    changed       = models.DateTimeField(null=True)
+    start_date    = models.DateField(auto_now=False, blank=True, null=True, auto_now_add=False)
+    end_date      = models.DateField(auto_now=False, blank=True, null=True, auto_now_add=False)
+    function      = models.CharField(max_length=15, blank=True, null=True, choices=FUNCTION_CHOICES, default='UNKNOWN')
+    learning_unit = models.ForeignKey('LearningUnit')
+    tutor         = models.ForeignKey('Tutor')
 
-handler404 = 'base.views.common.page_not_found'
-handler403 = 'base.views.common.access_denied'
-
-admin.site.site_header = 'OSIS'
-admin.site.site_title  = 'OSIS'
-admin.site.index_title = 'Louvain'
+    def __str__(self):
+        return u"%s - %s" % (self.tutor.person, self.function)
