+++ base/models/person.py
@@ -26,10 +26,8 @@
 
 
 from django.db import models
-from django.core.exceptions import ObjectDoesNotExist
 from django.contrib.auth.models import User
-from django.utils import timezone
-
+from django.utils.translation import ugettext_lazy as _
 
 class Person(models.Model):
     GENDER_CHOICES = (
