+++ base/models/tutor.py
@@ -1,18 +1,46 @@
+##############################################################################
+#
+#    OSIS stands for Open Student Information System. It's an application
+#    designed to manage the core business of higher education institutions,
+#    such as universities, faculties, institutes and professional schools.
+#    The core business involves the administration of students, teachers,
+#    courses, programs and so on.
+#
+#    Copyright (C) 2015-2016 Université catholique de Louvain (http://www.uclouvain.be)
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU General Public License as published by
+#    the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU General Public License for more details.
+#
+#    A copy of this license - GNU General Public License - is available
+#    at the root of the source code of this program.  If not,
+#    see http://www.gnu.org/licenses/.
+#
+##############################################################################
+from django.core.exceptions import ObjectDoesNotExist
+from django.db import models
+
+from base.models.person import find_person_by_user
+
+
 class Tutor(models.Model):
     external_id = models.CharField(max_length=100, blank=True, null=True)
     changed     = models.DateTimeField(null=True)
-    person      = models.ForeignKey(Person)
-
-    @staticmethod
-
+    person      = models.ForeignKey('Person')
 
     def __str__(self):
         return u"%s" % self.person
 
 
-def find_by_user(user):
+def find_tutor_by_user(user):
     try:
-        person = Person.find_person_by_user(user)
+        person = find_person_by_user(user)
         # tutor = Tutor.objects.filter(person=person)
         tutor = Tutor.objects.get(person = person)
