+++ base/templates/structure.html
@@ -103,13 +103,13 @@
         <div id="tab_offers" class="tab-pane fade">
           <div class="row">
             <div class="col-md-12">
-              {% if structure.find_offer_years_by_academic_year|length > 0%}
+              {% if structure.find_offer_years|length > 0%}
               <table class="table table-striped table-hover">
                 <thead>
                   <th>{% trans 'Acronym'%}</th>
                   <th>{% trans 'Title'%}</th>
                 </thead>
-                {% for oy in structure.find_offer_years_by_academic_year %}
+                {% for oy in structure.find_offer_years %}
                   {% ifchanged oy.academic_year %}
                        <tr style="background-color:#E5F2FF"><td colspan="2"><h4>{{oy.academic_year}}</h4></td></tr>
                   {% endifchanged %}
