+++ base/tests/util.py
@@ -1,6 +1,6 @@
 ##############################################################################
 #
-#    OSIS stands for Open Student Information System. It's an application
+# OSIS stands for Open Student Information System. It's an application
 #    designed to manage the core business of higher education institutions,
 #    such as universities, faculties, institutes and professional schools.
 #    The core business involves the administration of students, teachers,
@@ -26,36 +26,37 @@
 """
 Utility file for Tests units
 """
-from datetime import date
 from django.contrib.auth.models import User
 import os
-from selenium.common.exceptions import NoSuchElementException
-from datetime import date
-from base.models import Person
-from backoffice.settings import SCREEN_SHOT_FOLDER, FIREFOX_PROFILE_PATH, EMAIL_FILE_PATH
 
-ADMIN_USER='admin_user'
-GRANTED_USER='granted_user'
-VALID_USER='valid_user'
-INVALID_USER='invalid_user'
-PASSWORD='password'
-EMAIL='user@osis.org'
+from base.models.person import Person
+
+
+ADMIN_USER = 'admin_user'
+GRANTED_USER = 'granted_user'
+VALID_USER = 'valid_user'
+INVALID_USER = 'invalid_user'
+PASSWORD = 'password'
+EMAIL = 'user@osis.org'
+
 
 def init_admin_user():
     """
     Initialise an Admin user
     """
-    user = User.objects.create_superuser(ADMIN_USER,EMAIL,PASSWORD,is_staff=True)
+    user = User.objects.create_superuser(ADMIN_USER, EMAIL, PASSWORD, is_staff=True)
     user.save()
 
-def init_granted_user() :
+
+def init_granted_user():
     """
     Initialise a user tha has been granted to acces tested methods.
     If a new authorisation is set for a tested method , it has to be added to the "set_permissions_to_granted" methods.
     """
-    user = User.objects.create_user(GRANTED_USER,EMAIL,PASSWORD)
+    user = User.objects.create_user(GRANTED_USER, EMAIL, PASSWORD)
     set_permissions_to_granted(user)
     user.save()
+
 
 def set_permissions_to_granted(user):
     """
@@ -64,14 +65,16 @@ def set_permissions_to_granted(user):
     """
     return
 
+
 def init_valid_user():
     """
     Initialise a user that can log in , but doesn't have authorisations to access methods
     """
-    user = User.objects.create_user(VALID_USER,EMAIL,PASSWORD)
+    user = User.objects.create_user(VALID_USER, EMAIL, PASSWORD)
     user.save()
 
-def init_all_test_users() :
+
+def init_all_test_users():
     """
     Initialise all user types
     """
@@ -79,14 +82,16 @@ def init_all_test_users() :
     init_granted_user()
     init_valid_user()
 
-def email_destination_person() :
-    return Person(last_name='Test user',gender='M',email='gaetan.lamarca@uclouvain.be')
 
-def test_if_file_starting_with_exists(starting_with,file_dir_path):
+def email_destination_person():
+    return Person(last_name='Test user', gender='M', email='gaetan.lamarca@uclouvain.be')
+
+
+def test_if_file_starting_with_exists(starting_with, file_dir_path):
     filenames = os.listdir(file_dir_path)
     for filename in filenames:
-        if os.path.isfile(filename) and filename.startswith(starting_with) :
+        if os.path.isfile(filename) and filename.startswith(starting_with):
             return True
-    else :
+    else:
         return False
