+++ base/utils/upload_xls_utils.py
@@ -29,9 +29,18 @@ from django.core.urlresolvers import reverse
 from django.utils.translation import ungettext
 from openpyxl import load_workbook
 from django.contrib import messages
+from django.utils.translation import ugettext_lazy as _
+from base.enums import JUSTIFICATION_TYPES
 
-from base.models import *
 from base.forms import ScoreFileForm
+from base.models.academic_year import AcademicYear
+from base.models.exam_enrollment import ExamEnrollment
+from base.models.exam_enrollment_history import ExamEnrollmentHistory, create_exam_enrollment_historic
+from base.models.learning_unit_enrollment import LearningUnitEnrollment
+from base.models.learning_unit_year import LearningUnitYear
+from base.models.offer_enrollment import OfferEnrollment
+from base.models.offer_year import OfferYear
+from base.models.student import Student
 from base.utils import export_utils
 
 
@@ -153,7 +162,7 @@ def __save_xls_scores(request, file_name):
                                             exam_enrollment.score_draft = note
                                             exam_enrollment.justification_draft = justification_xls
                                             exam_enrollment.save()
-                                            ExamEnrollmentHistory.exam_enrollment_historic(request.user,exam_enrollment,note,justification_xls)
+                                            create_exam_enrollment_historic(request.user,exam_enrollment,note,justification_xls)
             data_line_number=data_line_number+1
         else:
             #Il faut valider le fichier xls
