+++ base/views/institution.py
@@ -25,7 +25,8 @@
 ##############################################################################
 from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
-from base.models import Structure
+
+from base.models.structure import find_structures, find_structure_by_id
 
 
 @login_required
@@ -41,7 +42,7 @@ def structures_search(request):
     acronym = request.GET['acronym']
     title = request.GET['title']
 
-    query = Structure.find_structures()
+    query = find_structures()
 
     if not acronym is None and len(acronym) > 0  :
         query = query.filter(acronym__icontains=acronym)
@@ -55,5 +56,5 @@ def structures_search(request):
 
 
 def structure_read(request,id):
-    structure = Structure.find_by_id(id)
+    structure = find_structure_by_id(id)
     return render(request, "structure.html", {'structure': structure})
