+++ base/views/learning_unit.py
@@ -1,6 +1,6 @@
 ##############################################################################
 #
-#    OSIS stands for Open Student Information System. It's an application
+# OSIS stands for Open Student Information System. It's an application
 #    designed to manage the core business of higher education institutions,
 #    such as universities, faculties, institutes and professional schools.
 #    The core business involves the administration of students, teachers,
@@ -25,24 +25,27 @@
 ##############################################################################
 from django.shortcuts import render
 
-from base.models import *
+from base.models.academic_year import find_academic_years
+from base.models.academic_year_calendar import current_academic_year
+from base.models.learning_unit_year import find_learning_unit_year_by_id, find_learning_unit_years_by_academic_year
+
 
 def learning_units(request):
-
     academic_year = None
     code = ""
 
-    academic_years = AcademicYear.find_academic_years()
-    academic_year_calendar = AcademicCalendar.current_academic_year()
+    academic_years = find_academic_years()
+    academic_year_calendar = current_academic_year()
 
     if not academic_year_calendar is None:
         academic_year = academic_year_calendar.id
     return render(request, "learning_units.html", {
-                                           'academic_year':  academic_year,
-                                           'code':           code,
-                                           'academic_years': academic_years,
-                                           'learning_units': [] ,
-                                           'init':           "1"})
+        'academic_year': academic_year,
+        'code': code,
+        'academic_years': academic_years,
+        'learning_units': [],
+        'init': "1"})
+
 
 def learning_units_search(request):
     """
@@ -52,24 +55,25 @@ def learning_units_search(request):
     academic_year = request.GET['academic_year']
     code = request.GET['code']
 
-    academic_years = AcademicYear.find_academic_years()
+    academic_years = find_academic_years()
 
     if academic_year is None:
-        academic_year_calendar = AcademicCalendar.current_academic_year()
+        academic_year_calendar = current_academic_year()
         if not academic_year_calendar is None:
             academic_year = academic_year_calendar.id
 
-    learning_units = LearningUnitYear.find_learning_unit_years_by_academic_year(academic_year)
+    learning_units = find_learning_unit_years_by_academic_year(academic_year)
 
-    if not code is None and len(code) > 0  :
+    if not code is None and len(code) > 0:
         learning_units = learning_units.filter(acronym__icontains=code)
 
-    return render(request, "learning_units.html", {'academic_year':  int(academic_year),
-                                           'code':           code,
-                                           'academic_years': academic_years,
-                                           'learning_units': learning_units ,
-                                           'init':           "0"})
+    return render(request, "learning_units.html", {'academic_year': int(academic_year),
+                                                   'code': code,
+                                                   'academic_years': academic_years,
+                                                   'learning_units': learning_units,
+                                                   'init': "0"})
 
-def learning_unit_read(request,learning_unit_id):
-    learning_unit_year = LearningUnitYear.find_learning_unit_year_by_id(learning_unit_id)
-    return render(request, "learning_unit.html", {'learning_unit_year':    learning_unit_year})
+
+def learning_unit_read(request, learning_unit_id):
+    learning_unit_year = find_learning_unit_year_by_id(learning_unit_id)
+    return render(request, "learning_unit.html", {'learning_unit_year': learning_unit_year})
