+++ base/views/score_encoding.py
@@ -1,6 +1,6 @@
 ##############################################################################
 #
-#    OSIS stands for Open Student Information System. It's an application
+# OSIS stands for Open Student Information System. It's an application
 #    designed to manage the core business of higher education institutions,
 #    such as universities, faculties, institutes and professional schools.
 #    The core business involves the administration of students, teachers,
@@ -27,37 +27,48 @@ from django.core.urlresolvers import reverse
 from django.http import HttpResponseRedirect
 from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
-from base.utils import send_mail,pdf_utils,export_utils
-from base.models import *
+
+from base.enums import JUSTIFICATION_TYPES
+
+from base.models.academic_year_calendar import current_academic_year
+from base.models.attribution import Attribution
+from base.models.exam_enrollment import calculate_exam_enrollment_progress, \
+    find_exam_enrollments_by_session, find_exam_enrollments_drafts_by_session, count_encoded_scores, \
+    find_exam_enrollments_to_validate_by_session
+from base.models.exam_enrollment_history import ExamEnrollmentHistory, create_exam_enrollment_historic
+from base.models.program_manager import find_faculty_by_user
+from base.models.session_exam import find_session_by_id, find_sessions_by_tutor, find_sessions_by_faculty
+from base.models.tutor import find_tutor_by_user
+from base.utils import send_mail, pdf_utils, export_utils
 
 
 @login_required
 def scores_encoding(request):
-    academic_year = AcademicCalendar.current_academic_year()
+    academic_year = current_academic_year()
 
-    tutor = Tutor.find_by_user(request.user)
+    tutor = find_tutor_by_user(request.user)
     # In case the user is a tutor.
     sessions = None
     faculty = None
     if tutor:
-        sessions = SessionExam.find_sessions_by_tutor(tutor, academic_year)
+        sessions = find_sessions_by_tutor(tutor, academic_year)
     # In case the user is not a tutor we check whether it is member of a faculty.
     elif request.user.groups.filter(name='FAC').exists():
-        faculty = ProgrammeManager.find_faculty_by_user(request.user)
+        faculty = find_faculty_by_user(request.user)
         if faculty:
-            sessions = SessionExam.find_sessions_by_faculty(faculty, academic_year)
+            sessions = find_sessions_by_faculty(faculty, academic_year)
 
     # Calculate the progress of all courses of the tutor.
     all_enrollments = []
     session = None
     if sessions:
         for session in sessions:
-            enrollments = list(ExamEnrollment.find_exam_enrollments(session))
+            enrollments = list(find_exam_enrollments_by_session(session))
             if enrollments:
                 all_enrollments = all_enrollments + enrollments
         session = sessions.first()
 
-    progress = ExamEnrollment.calculate_progress(all_enrollments)
+    progress = calculate_exam_enrollment_progress(all_enrollments)
 
     return render(request, "scores_encoding.html",
                   {'section': 'scores_encoding',
@@ -72,15 +83,15 @@ def scores_encoding(request):
 @login_required
 def online_encoding(request, session_id):
     tutor = None
-    faculty = ProgrammeManager.find_faculty_by_user(request.user)
+    faculty = find_faculty_by_user(request.user)
     if not faculty:
-        tutor = Tutor.find_by_user(request.user)
+        tutor = find_tutor_by_user(request.user)
 
-    academic_year = AcademicCalendar.current_academic_year()
-    session = SessionExam.find_session(session_id)
-    enrollments = ExamEnrollment.find_exam_enrollments(session)
-    progress = ExamEnrollment.calculate_progress(enrollments)
-    num_encoded_scores = ExamEnrollment.count_encoded_scores(enrollments)
+    academic_year = current_academic_year()
+    session = find_session_by_id(session_id)
+    enrollments = find_exam_enrollments_by_session(session)
+    progress = calculate_exam_enrollment_progress(enrollments)
+    num_encoded_scores = count_encoded_scores(enrollments)
 
     return render(request, "online_encoding.html",
                   {'section': 'scores_encoding',
@@ -95,11 +106,11 @@ def online_encoding(request, session_id):
 
 @login_required
 def online_encoding_form(request, session_id):
-    session = SessionExam.find_session(session_id)
-    enrollments = ExamEnrollment.find_exam_enrollments(session)
+    session = find_session_by_id(session_id)
+    enrollments = find_exam_enrollments_by_session(session)
     if request.method == 'GET':
-        tutor = Tutor.find_by_user(request.user)
-        academic_year = AcademicCalendar.current_academic_year()
+        tutor = find_tutor_by_user(request.user)
+        academic_year = current_academic_year()
         return render(request, "online_encoding_form.html",
                       {'section': 'scores_encoding',
                        'tutor': tutor,
@@ -127,11 +138,11 @@ def online_encoding_form(request, session_id):
 
 @login_required
 def online_double_encoding_form(request, session_id):
-    session = SessionExam.find_session(session_id)
-    enrollments = ExamEnrollment.find_draft_exam_enrollments(session)
+    session = find_session_by_id(session_id)
+    enrollments = find_exam_enrollments_drafts_by_session(session)
     if request.method == 'GET':
-        tutor = Tutor.find_by_user(request.user)
-        academic_year = AcademicCalendar.current_academic_year()
+        tutor = find_tutor_by_user(request.user)
+        academic_year = current_academic_year()
         return render(request, "online_double_encoding_form.html",
                       {'section': 'scores_encoding',
                        'tutor': tutor,
@@ -163,11 +174,11 @@ def online_double_encoding_form(request, session_id):
 
 @login_required
 def online_double_encoding_validation(request, session_id):
-    session_exam = SessionExam.find_session(session_id)
+    session_exam = find_session_by_id(session_id)
     if request.method == 'GET':
-        tutor = Tutor.find_by_user(request.user)
-        academic_year = AcademicCalendar.current_academic_year()
-        enrollments = ExamEnrollment.find_exam_enrollments_to_validate(session_exam)
+        tutor = find_tutor_by_user(request.user)
+        academic_year = current_academic_year()
+        enrollments = find_exam_enrollments_to_validate_by_session(session_exam)
         return render(request, "online_double_encoding_validation.html",
                       {'section': 'scores_encoding',
                        'tutor': tutor,
@@ -177,7 +188,7 @@ def online_double_encoding_validation(request, session_id):
                        'justifications': JUSTIFICATION_TYPES})
 
     elif request.method == 'POST':
-        enrollments = ExamEnrollment.find_exam_enrollments(session_exam)
+        enrollments = find_exam_enrollments_by_session(session_exam)
 
         for enrollment in enrollments:
             score = request.POST.get('score_' + str(enrollment.id), None)
@@ -194,7 +205,7 @@ def online_double_encoding_validation(request, session_id):
                 enrollment.justification_final = justification
                 enrollment.justification_draft = enrollment.justification_final
             if score or justification:
-                ExamEnrollmentHistory.exam_enrollment_historic(request.user,enrollment,score,justification)
+                create_exam_enrollment_historic(request.user, enrollment, score, justification)
             enrollment.justification_reencoded = None
             enrollment.save()
 
@@ -212,8 +223,8 @@ def online_double_encoding_validation(request, session_id):
 
 @login_required
 def online_encoding_submission(request, session_id):
-    session_exam = SessionExam.find_session(session_id)
-    enrollments = ExamEnrollment.find_draft_exam_enrollments(session_exam)
+    session_exam = find_session_by_id(session_id)
+    enrollments = find_exam_enrollments_drafts_by_session(session_exam)
     all_encoded = True
     for enrollment in enrollments:
         if enrollment.score_draft or enrollment.justification_draft:
@@ -222,7 +233,8 @@ def online_encoding_submission(request, session_id):
             if enrollment.justification_draft:
                 enrollment.justification_final = enrollment.justification_draft
             enrollment.save()
-            ExamEnrollmentHistory.exam_enrollment_historic(request.user,enrollment,enrollment.score_final,enrollment.justification_final)
+            create_exam_enrollment_historic(request.user, enrollment, enrollment.score_final,
+                                                           enrollment.justification_final)
         else:
             all_encoded = False
 
@@ -234,18 +246,18 @@ def online_encoding_submission(request, session_id):
     learning_unit = session_exam.learning_unit_year.learning_unit
     attributions = Attribution.objects.filter(learning_unit=learning_unit)
     persons = [attribution.tutor.person for attribution in attributions if attribution.function == 'PROFESSOR']
-    send_mail.send_mail_after_scores_submission(persons,learning_unit.acronym)
+    send_mail.send_mail_after_scores_submission(persons, learning_unit.acronym)
 
     return HttpResponseRedirect(reverse('online_encoding', args=(session_id,)))
 
 
 @login_required
-def notes_printing(request,session_exam_id,learning_unit_year_id):
-    tutor = Tutor.find_by_user(request.user)
-    academic_year = AcademicCalendar.current_academic_year()
-    session_exam = SessionExam.find_session(session_exam_id)
-    sessions = SessionExam.find_sessions_by_tutor(tutor, academic_year)
-    return pdf_utils.print_notes(request,tutor,academic_year,session_exam,sessions,learning_unit_year_id)
+def notes_printing(request, session_exam_id, learning_unit_year_id):
+    tutor = find_tutor_by_user(request.user)
+    academic_year = current_academic_year()
+    session_exam = find_session_by_id(session_exam_id)
+    sessions = find_sessions_by_tutor(tutor, academic_year)
+    return pdf_utils.print_notes(request, tutor, academic_year, session_exam, sessions, learning_unit_year_id)
 
 
 @login_required
@@ -255,18 +267,19 @@ def upload_score_error(request):
 
 @login_required
 def notes_printing(request, session_exam_id, learning_unit_year_id):
-    tutor = Tutor.find_by_user(request.user)
-    academic_year = AcademicCalendar.current_academic_year()
-    session_exam = SessionExam.find_session(session_exam_id)
-    sessions = SessionExam.find_sessions_by_tutor(tutor, academic_year)
-    return pdf_utils.print_notes(request,tutor,academic_year,session_exam,sessions,learning_unit_year_id)
+    tutor = find_tutor_by_user(request.user)
+    academic_year = current_academic_year()
+    session_exam = find_session_by_id(session_exam_id)
+    sessions = find_sessions_by_tutor(tutor, academic_year)
+    return pdf_utils.print_notes(request, tutor, academic_year, session_exam, sessions, learning_unit_year_id)
 
 
 @login_required
 def notes_printing_all(request, session_id):
-    return notes_printing(request,session_id,-1)
+    return notes_printing(request, session_id, -1)
 
 
 @login_required
 def export_xls(request, session_id, learning_unit_year_id, academic_year_id):
-    return export_utils.export_xls(request, session_id, learning_unit_year_id, academic_year_id, request.user.groups.filter(name='FAC').exists())
+    return export_utils.export_xls(request, session_id, learning_unit_year_id, academic_year_id,
+                                   request.user.groups.filter(name='FAC').exists())
